Day one plus workspaces
I really did poorly today, couldn't find a way to a solution that didn't
involve several weird loops of loops of loops. Fortunately this isn't a
contest, and looking at others' code for help and ideas (to learn) is
not a bad approach. Wow did I miss some obvious methods.